container and crafting item handling
campfire extinguishers/ignitors didn't work correctly for the kinds of items that store crafting durability in nbt and use container items to damage themselves.
it should be fixed now for stuff like gregtech tools. for example, a flint sword now works properly via the default config ignitor "left/[ench:20,1]/damageable".
previously, you could work around this by manually adding a "stackable" rule, but that's pretty counter-intuitive.

also, items that get returned as outputs are added to the inventory in the normal way instead of in a "smart" way.